Git status line in git repositories.
When inside a git repository, the checked-out branch is now displayed automatically.
Also, if there are staged but uncommitted changes, a blue dot is displayed. If there are unstaged changes, a red one is.

This does not play nice with long input lines. I know how to fix this, but right now it's too much effort to.

This is vaguely inspired by http://briancarper.net/blog/570/git-info-in-your-zsh-prompt and the peoples in the comments, except simpler, because this is all the information I need, mostly.