Status bar Gmail notifier now updates after closing mutt.
Duplicated code from wmiirc, but it's not worth spinning off into a separate script.
Also a shame about the namespace pollution, but it's the best way I see that also still allows you to use plain mutt if needed and to pass command line arguments to it even if you don't.